storage: move determine_sink_resume_upper to mz_storage
This is pure code movement. This function is only used in the dataflow
during sink initialization and so it is moved there.
